Moved content of index.jsx to new file App.jsx to keep naming of file to component consistent. Removed test line from README, and slightly modified routes/query.js in Mariner. `
